refactor: Use a base class to contain the common functions of the dual and state classes
