+++ src/serac/physics/utilities/finite_element_state.hpp
@@ -20,6 +20,7 @@
 #include "mfem.hpp"
 
 #include "serac/physics/utilities/variant.hpp"
+#include "serac/physics/utilities/finite_element_vector.hpp"
 
 namespace serac {
 
@@ -45,57 +46,22 @@ inline bool is_vector_valued(const GeneralCoefficient& coef)
 }
 
 /**
- * @brief Class for encapsulating the critical MFEM components of a solver
+ * @brief Class for encapsulating the critical MFEM components of a primal finite element field
  *
  * Namely: Mesh, FiniteElementCollection, FiniteElementState,
  * GridFunction, and a Vector of the solution
  */
-class FiniteElementState {
+class FiniteElementState : public FiniteElementVector {
 public:
   /**
-   * @brief Structure for optionally configuring a FiniteElementState
-   */
-  // The optionals are explicitly default-constructed to allow the user to partially aggregrate-initialized
-  // with only the options they care about
-  struct Options {
-    /**
-     * @brief The polynomial order that should be used for the problem
-     */
-    int order = 1;
-    /**
-     * @brief The number of copies of the finite element collections (e.g. vector_dim = 2 or 3 for solid mechanics).
-     * Defaults to scalar valued spaces.
-     */
-    int vector_dim = 1;
-    /**
-     * @brief The FECollection to use - defaults to an H1_FECollection
-     */
-    std::unique_ptr<mfem::FiniteElementCollection> coll = {};
-    /**
-     * The DOF ordering that should be used interally by MFEM
-     */
-    mfem::Ordering::Type ordering = mfem::Ordering::byVDIM;
-    /**
-     * @brief The name of the field encapsulated by the state object
-     */
-    std::string name = "";
-    /**
-     * @brief Whether the GridFunction should be allocated (and owned by the FEState object)
-     */
-    bool alloc_gf = true;
-  };
-
-  /**
-   * Main constructor for building a new state object
+   * Main constructor for building a new finite element state
    * @param[in] mesh The problem mesh (object does not take ownership)
    * @param[in] options The options specified, namely those relating to the order of the problem,
    * the dimension of the FESpace, the type of FEColl, the DOF ordering that should be used,
    * and the name of the field
    */
-  FiniteElementState(
-      mfem::ParMesh& mesh,
-      Options&&      options = {
-          .order = 1, .vector_dim = 1, .coll = {}, .ordering = mfem::Ordering::byVDIM, .name = "", .alloc_gf = true});
+  FiniteElementState(mfem::ParMesh& mesh, Options&& options = {})
+      : FiniteElementVector(mesh, std::forward<FiniteElementVector::Options>(options)){};
 
   /**
    * @brief Minimal constructor for a FiniteElementState given an already-existing field
@@ -103,7 +69,8 @@ public:
    * @param[in] gf The field for the state to create (object does not take ownership)
    * @param[in] name The name of the field
    */
-  FiniteElementState(mfem::ParMesh& mesh, mfem::ParGridFunction& gf, const std::string& name = "");
+  FiniteElementState(mfem::ParMesh& mesh, mfem::ParGridFunction& gf, const std::string& name = "")
+      : FiniteElementVector(mesh, gf, name){};
 
   /**
    * @brief Minimal constructor for a FiniteElementState given a finite element space
@@ -111,7 +78,8 @@ public:
    * @param[in] space The space to use for the finite element state. This space is deep copied into the new FE state
    * @param[in] name The name of the field
    */
-  FiniteElementState(mfem::ParMesh& mesh, mfem::ParFiniteElementSpace& space, const std::string& name = "");
+  FiniteElementState(mfem::ParMesh& mesh, mfem::ParFiniteElementSpace& space, const std::string& name = "")
+      : FiniteElementVector(mesh, space, name){};
 
   /**
    * @brief Minimal constructor for a FiniteElementState given an already-existing state
@@ -119,12 +87,8 @@ public:
    * @param[in] fe_state The state for the new state to copy
    * @param[in] name The name of the field
    */
-  FiniteElementState(mfem::ParMesh& mesh, FiniteElementState& fe_state, const std::string& name = "");
-
-  /**
-   * Returns the MPI communicator for the state
-   */
-  MPI_Comm comm() const { return detail::retrieve(space_).GetComm(); }
+  FiniteElementState(mfem::ParMesh& mesh, FiniteElementState& fe_state, const std::string& name = "")
+      : FiniteElementVector(mesh, fe_state.gridFunc(), name){};
 
   /**
    * Returns a non-owning reference to the internal grid function
@@ -149,30 +113,6 @@ public:
   {
     return mfem::VectorGridFunctionCoefficient{&detail::retrieve(gf_)};
   }
-
-  /**
-   * Returns a non-owning reference to the internal mesh object
-   */
-  mfem::ParMesh& mesh() { return mesh_; }
-
-  /**
-   * Returns a non-owning reference to the internal FESpace
-   */
-  mfem::ParFiniteElementSpace& space() { return detail::retrieve(space_); }
-  /// \overload
-  const mfem::ParFiniteElementSpace& space() const { return detail::retrieve(space_); }
-
-  /**
-   * Returns a non-owning reference to the vector of true DOFs
-   */
-  mfem::HypreParVector& trueVec() { return true_vec_; }
-  /// \overload
-  const mfem::HypreParVector& trueVec() const { return true_vec_; }
-
-  /**
-   * Returns the name of the FEState (field)
-   */
-  std::string name() const { return name_; }
 
   /**
    * Projects a coefficient (vector or scalar) onto the field
@@ -201,93 +141,20 @@ public:
     detail::retrieve(gf_).ProjectCoefficient(coef);
     initializeTrueVec();
   }
-
-  /**
-   * Initialize the true DOF vector by extracting true DOFs from the internal
-   * grid function into the internal true DOF vector
-   */
-  void initializeTrueVec() { detail::retrieve(gf_).GetTrueDofs(true_vec_); }
-
-  /**
-   * Set the internal grid function using the true DOF values
-   */
-  void distributeSharedDofs() { detail::retrieve(gf_).SetFromTrueDofs(true_vec_); }
-
   /**
    * @brief Set a finite element state to a constant value
    *
-   * @param value The constant to set the finite element state to
-   * @return The modified finite element state
+   * @param value The constant to set the finite element dual to
+   * @return The modified finite element dual
    * @note This sets the true degrees of freedom and then broadcasts to the shared grid function entries. This means
    * that if a different value is given on different processors, a shared DOF will be set to the owning processor value.
    */
-  FiniteElementState& operator=(const double value);
-
-  /**
-   * Utility function for creating a tensor, e.g. mfem::HypreParVector,
-   * mfem::ParBilinearForm, etc on the FESpace encapsulated by an FEState object
-   * @return An owning pointer to a heap-allocated tensor
-   * @pre Tensor must have the constructor Tensor::Tensor(ParFiniteElementSpace*)
-   */
-  template <typename Tensor>
-  std::unique_ptr<Tensor> createOnSpace()
+  FiniteElementState& operator=(const double value)
   {
-    static_assert(std::is_constructible_v<Tensor, mfem::ParFiniteElementSpace*>,
-                  "Tensor must be constructible with a ptr to ParFESpace");
-    return std::make_unique<Tensor>(&detail::retrieve(space_));
+    FiniteElementVector::operator=(value);
+    return *this;
   }
-
-private:
-  /**
-   * @brief A reference to the mesh object on which the field is defined
-   */
-  std::reference_wrapper<mfem::ParMesh> mesh_;
-  /**
-   * @brief Possibly-owning handle to the FiniteElementCollection, as it is owned
-   * by the FiniteElementState in a normal run and by the MFEMSidreDataCollection
-   * in a restart run
-   * @note Must be const as FESpaces store a const reference to their FEColls
-   */
-  detail::MaybeOwningPointer<const mfem::FiniteElementCollection> coll_;
-  /**
-   * @brief Possibly-owning handle to the FiniteElementCollection, as it is owned
-   * by the FiniteElementState in a normal run and by the MFEMSidreDataCollection
-   * in a restart run
-   */
-  detail::MaybeOwningPointer<mfem::ParFiniteElementSpace> space_;
-  /**
-   * @brief Possibly-owning handle to the ParGridFunction, as it is owned
-   * by the FiniteElementState in a normal run and by the MFEMSidreDataCollection
-   * in a restart run
-   */
-  detail::MaybeOwningPointer<mfem::ParGridFunction> gf_;
-  mfem::HypreParVector                              true_vec_;
-  std::string                                       name_ = "";
 };
-
-/**
- * @brief Find the average value of a finite element state across all nodes
- *
- * @param state The state variable to compute a max of
- * @return The average value
- */
-double avg(const FiniteElementState& state);
-
-/**
- * @brief Find the max value of a finite element state across all nodes
- *
- * @param state The state variable to compute a max of
- * @return The max value
- */
-double max(const FiniteElementState& state);
-
-/**
- * @brief Find the min value of a finite element state across all nodes
- *
- * @param state The state variable to compute a min of
- * @return The min value
- */
-double min(const FiniteElementState& state);
 
 /**
  * @brief Calculate the Lp norm of a finite element state
