+++ src/serac/physics/utilities/state_manager.cpp
@@ -63,7 +63,7 @@ void StateManager::initialize(axom::sidre::DataStore& ds, const std::string& col
   }
 }
 
-FiniteElementState StateManager::newState(FiniteElementState::Options&& options)
+FiniteElementState StateManager::newState(FiniteElementVector::Options&& options)
 {
   SLIC_ERROR_ROOT_IF(!datacoll_, "Serac's datacollection was not initialized - call StateManager::initialize first");
   const std::string name = options.name;
@@ -73,7 +73,7 @@ FiniteElementState StateManager::newState(FiniteElementState::Options&& options)
   } else {
     SLIC_ERROR_ROOT_IF(datacoll_->HasField(name),
                        fmt::format("Serac's datacollection was already given a field named '{0}'", name));
-    options.alloc_gf = false;
+    options.alloc_local = false;
     FiniteElementState state(mesh(), std::move(options));
     datacoll_->RegisterField(name, &(state.gridFunc()));
     // Now that it's been allocated, we can set it to zero
@@ -82,7 +82,7 @@ FiniteElementState StateManager::newState(FiniteElementState::Options&& options)
   }
 }
 
-FiniteElementDual StateManager::newDual(FiniteElementDual::Options&& options)
+FiniteElementDual StateManager::newDual(FiniteElementVector::Options&& options)
 {
   SLIC_ERROR_ROOT_IF(!datacoll_, "Serac's datacollection was not initialized - call StateManager::initialize first");
   const std::string name = options.name;
